ルーキー参加曲一覧のSSG化
