+++ 2026spring/frontend/src/app/submissions/songs/rookie/page.tsx
@@ -1,287 +1,21 @@
-"use client"
-
-import { useEffect, useState, useRef } from "react"
-import Image from "next/image"
+import VideoList from "@/components/video/VideoList"
 import { CONFIG } from "@/config/config"
-import { getCurrentPhase, EVENT_PHASES } from "@/config/phase"
-import TBA from "@/components/TBA"
+import { fetchVideosSheet } from "@/lib/fetchSheet"
 
-type Video = {
-  title: string
-  author: string
-  videoUrl: string
-  thumbnailUrl: string
-  publishedAt?: string
-  description?: string
-}
 
-const VIEW_PHASE = {
-  BEFORE: "before",
-  DURING: "during",
-} as const
-
-function getViewPhase(phase: string) {
-  switch (phase) {
-    case EVENT_PHASES.BEFORE:
-    case EVENT_PHASES.OPENING:
-      return VIEW_PHASE.BEFORE
-    default:
-      return VIEW_PHASE.DURING
-  }
-}
-
-function parseDate(dateStr?: string) {
-  if (!dateStr) return 0
-  return new Date(dateStr).getTime()
-}
-
-function shuffleArray<T>(array: T[]): T[] {
-  const arr = [...array]
-  for (let i = arr.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1))
-    ;[arr[i], arr[j]] = [arr[j], arr[i]]
-  }
-  return arr
-}
-
-const PAGE_SIZE = 24
-
-export default function Page() {
-  const [data, setData] = useState<Video[]>([])
-  const [displayData, setDisplayData] = useState<Video[]>([])
-  const [visibleCount, setVisibleCount] = useState(PAGE_SIZE)
-
-  const [loading, setLoading] = useState(true)
-  const [searchText, setSearchText] = useState("")
-  const [sortType, setSortType] = useState<"new" | "old">("new")
-  const [isRandom, setIsRandom] = useState(false)
-
-  const loadMoreRef = useRef<HTMLDivElement | null>(null)
-
-  const phase = getCurrentPhase()
-  const viewPhase = getViewPhase(phase)
-
-  /* 初期ロード */
-  useEffect(() => {
-    fetch("/api/submissions/songs/rookie")
-      .then(res => res.json())
-      .then(res => {
-        const mapped = res.map((item: any) => ({
-          title: item.title,
-          author: item.creator,
-          videoUrl: item.videoUrl,
-          thumbnailUrl: item.thumbnailUrl,
-          publishedAt: item.publishedAt,
-          description: item.description
-        }))
-        setData(mapped)
-        setLoading(false)
-      })
-  }, [])
-
-  /* フィルタ + ソート */
-  useEffect(() => {
-    if (data.length === 0) return
-    if (isRandom) return
-
-    let filtered = [...data]
-
-    if (searchText.trim()) {
-      const q = searchText.toLowerCase()
-      filtered = filtered.filter(v =>
-        (v.title + v.author).toLowerCase().includes(q)
-      )
-    }
-
-    if (sortType === "new") {
-      filtered.sort((a, b) => parseDate(b.publishedAt) - parseDate(a.publishedAt))
-    } else {
-      filtered.sort((a, b) => parseDate(a.publishedAt) - parseDate(b.publishedAt))
-    }
-
-    setDisplayData(filtered)
-    setVisibleCount(PAGE_SIZE)
-  }, [data, searchText, sortType, isRandom])
-
-  /* ランダム */
-  const handleShuffle = () => {
-    let base = [...data]
-
-    if (searchText.trim()) {
-      const q = searchText.toLowerCase()
-      base = base.filter(v =>
-        (v.title + v.author).toLowerCase().includes(q)
-      )
-    }
-
-    setDisplayData(shuffleArray(base))
-    setIsRandom(true)
-    setVisibleCount(PAGE_SIZE)
-  }
-
-  const handleSortChange = (val: "new" | "old") => {
-    setSortType(val)
-    setIsRandom(false)
-  }
-
-  const handleSearchChange = (val: string) => {
-    setSearchText(val)
-    setIsRandom(false)
-  }
-
-  /* 無限スクロール */
-  useEffect(() => {
-    if (!loadMoreRef.current) return
-
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting) {
-          setVisibleCount(prev => prev + PAGE_SIZE)
-        }
-      },
-      { rootMargin: "200px" }
-    )
-
-    observer.observe(loadMoreRef.current)
-    return () => observer.disconnect()
-  }, [displayData])
-
-  if (viewPhase === VIEW_PHASE.BEFORE) {
-    return <TBA title={`楽曲一覧 ルーキー`} />
-  }
-
-  const visibleItems = displayData.slice(0, visibleCount)
-
-  return (
-    <main className="flex justify-center">
-      <div className="w-full max-w-6xl px-4 py-6">
-
-        {/* ヘッダー */}
-        <div className="text-center mb-8">
-          <h1 className="text-3xl md:text-4xl font-bold leading-tight">
-            楽曲一覧 ルーキー
-          </h1>
-          <p className="text-sm text-gray-500 mt-1">
-            「{CONFIG.event.name}」のルーキー参加楽曲を掲載しています。
-          </p>
-          <div className="mt-4 border-b border-gray-200 w-full" />
-        </div>
-
-        {/* 操作バー */}
-        {!loading && (
-          <div className="flex flex-col gap-3 mb-4">
-            <div className="flex justify-start">
-              <input
-                type="text"
-                placeholder="検索（タイトル・投稿者）"
-                value={searchText}
-                onChange={(e) => handleSearchChange(e.target.value)}
-                className="w-full sm:w-72 border rounded px-3 py-1.5 text-sm"
-              />
-            </div>
-
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <select
-                  value={isRandom ? "random" : sortType}
-                  onChange={(e) => {
-                    const val = e.target.value
-                    if (val === "random") return
-                    handleSortChange(val as "new" | "old")
-                  }}
-                  className="border rounded px-2 py-1 text-sm"
-                >
-                  {isRandom && <option value="random">ランダム</option>}
-                  <option value="new">新しい順</option>
-                  <option value="old">古い順</option>
-                </select>
-
-                <button
-                  onClick={handleShuffle}
-                  className="
-                    border rounded
-                    px-3 py-1
-                    text-sm
-                    bg-blue-50 text-blue-700
-                    hover:bg-blue-100
-                    font-semibold
-                    active:scale-95 transition
-                  "
-                >
-                  ランダムに並び替え
-                </button>
-              </div>
-
-              <div className="text-sm text-gray-600">
-                {displayData.length}件
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* スケルトン（完全維持） */}
-        {loading && (
-          <div className="grid grid-cols-2 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 sm:gap-6">
-            {Array.from({ length: 8 }).map((_, i) => (
-              <div key={i} className="flex flex-col border border-gray-200 rounded-xl overflow-hidden">
-                <div className="aspect-video bg-gray-200 animate-pulse" />
-                <div className="p-2 space-y-2">
-                  <div className="h-3 w-24 bg-gray-200 rounded animate-pulse" />
-                  <div className="h-4 bg-gray-200 rounded animate-pulse" />
-                  <div className="h-3 w-1/2 bg-gray-200 rounded animate-pulse" />
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-
-        {/* 空 */}
-        {!loading && displayData.length === 0 && (
-          <div className="text-center py-20 text-gray-600">
-            ルーキー楽曲はまだありません。
-          </div>
-        )}
-
-        {/* グリッド */}
-        {!loading && displayData.length > 0 && (
-          <>
-            <div className="grid grid-cols-2 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 sm:gap-6">
-              {visibleItems.map((video, i) => (
-                <div key={i} className="flex flex-col group border border-gray-200 rounded-xl overflow-hidden transition hover:shadow-md">
-                  <a href={video.videoUrl} target="_blank" className="flex flex-col">
-
-                    <div className="relative aspect-video overflow-hidden">
-                      <Image
-                        src={video.thumbnailUrl}
-                        alt={video.title}
-                        fill
-                        sizes="(max-width: 768px) 50vw, (max-width: 1200px) 33vw, 25vw"
-                        className="object-cover transition group-hover:scale-105"
-                        loading="lazy"
-                        unoptimized
-                      />
-                    </div>
-
-                    <div className="flex flex-col mt-2 px-2 pb-2">
-                      <h2 className="text-sm font-bold leading-snug line-clamp-2 min-h-[2.8rem] group-hover:underline">
-                        {video.title}
-                      </h2>
-                      <p className="text-xs text-gray-600 mt-1 truncate">
-                        {video.author}
-                      </p>
-                    </div>
-
-                  </a>
-                </div>
-              ))}
-            </div>
-
-            {/* 無限スクロール */}
-            <div ref={loadMoreRef} className="h-10" />
-          </>
-        )}
-
-      </div>
-    </main>
+export default async function Page() {
+  const rawVideos = await fetchVideosSheet(
+    CONFIG.videosheets.rookie.name
   )
+
+  const mappedVideos = rawVideos.map((item) => ({
+    title: item.title,
+    author: item.creator, // creator を author にマッピング
+    videoUrl: item.videoUrl,
+    thumbnailUrl: item.thumbnailUrl,
+    publishedAt: item.publishedAt,
+    description: item.description
+  }))
+
+  return <VideoList initialData={mappedVideos} />
 }
